Moved samples to package
