--- Assets/Examples.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 620eb5da8cfff4953a436052073e7107
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
