+++ Packages/com.cemuka.runtime-node-editor/Samples~/ColorEditor/Cemuka.RuntimeNodeEditor.Samples.ColorEditor.asmdef
@@ -1,0 +1,17 @@
+{
+    "name": "Cemuka.RuntimeNodeEditor.Samples.ColorEditor",
+    "rootNamespace": "",
+    "references": [
+        "GUID:8f5c68f774363a54c99fc0044d54659f",
+        "GUID:6055be8ebefd69e48b49212b09b47b2f"
+    ],
+    "includePlatforms": [],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": false,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
+}
