+++ Packages/com.cemuka.runtime-node-editor/Samples~/ColorEditor/Resources/Nodes/ColorDisplay.prefab.meta
