+++ Packages/com.cemuka.runtime-node-editor/Samples~/ColorEditor/Scripts/ColorDisplayNode.cs
