+++ Packages/com.cemuka.runtime-node-editor/Samples~/ColorEditor/Scripts/ColorEditorStarter.cs
