+++ Packages/com.cemuka.runtime-node-editor/Samples~/ColorEditor/Scripts/ColorNodeEditor.cs
