+++ Packages/com.cemuka.runtime-node-editor/Samples~/SimpleMathEditor/Cemuka.RuntimeNodeEditor.Samples.SimpleMathEditor.asmdef.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: f7012aefaa6b047228a8db4273497b4f
-folderAsset: yes
-DefaultImporter:
+guid: 527e36fae183661439c48746e00d41a2
+AssemblyDefinitionImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
