+++ Packages/com.cemuka.runtime-node-editor/Samples~/SimpleMathEditor/Scripts/ApplicationStartup.cs
