+++ Packages/com.cemuka.runtime-node-editor/Samples~/SimpleMathEditor/Scripts/MathOperationNode.cs.meta
